Place statement and operation logic in submodules
